ft: move app routing to a different module
