+++ src/app/app.module.ts
@@ -2,19 +2,9 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { RouterModule, Routes } from '@angular/router';
 
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-
-const appRoutes: Routes = [
-  { path: 'auth', component: null }, // signin and signup
-  { path: 'home', component: null }, // for viewing all the documents
-  { path: 'setting', component: null }, // user profiles
-  { path: 'documents/create', component: null}, // create a new document
-  { path: 'documents/:id', component: null}, // viewing a document
-  { path: 'documents/:id/edit', component: null}, // editing a document
-  { path: '**', component: null}, // for all other routes that don't exist
-];
 
 @NgModule({
   declarations: [
@@ -24,7 +14,7 @@ const appRoutes: Routes = [
     BrowserModule,
     FormsModule,
     HttpModule,
-    RouterModule.forRoot(appRoutes)
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]
